Broke out internal classes from SinglePoint
